src/auth/requirements.txt
astroid==2.14.1
click==8.1.3
dill==0.3.6
Flask==2.2.2
Flask-MySQL==1.5.2
Flask-MySQLdb==1.0.1
importlib-metadata==6.0.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mysqlclient==2.1.1
parso==0.8.3
platformdirs==2.6.2
PyJWT==2.6.0
pylint==2.16.0
PyMySQL==1.0.2
tomli==2.0.1
tomlkit==0.11.6
typing-extensions==4.4.0
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.12.0
